+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.2.0 (DEV)
 
+- Fix: Incorrect border in new MPX file with custom size.
 - Fix: Some text grammar.
 - Add: Flood fill drawing tool.
 - Add: Display hovered cell coordinates in footer.