+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 ## 0.2.0 (DEV)
 
-- Fix: Incorrect border in new MPX file with custom size.
-- Fix: Some text grammar.
 - Add: Flood fill drawing tool.
 - Add: Display hovered cell coordinates in footer.
 - Add: Open first level when a file opened or created.
+- Fix: Incorrect border in new MPX file with custom size.
+- Fix: Lost which levels was previously opened on page load.
+- Fix: Some text grammar.
 
 ## 0.1.0 (2022-10-13)