+++ CHANGELOG.md
@@ -5,8 +5,10 @@
 - Add: Flood fill drawing tool.
 - Add: Display hovered cell coordinates in footer.
 - Add: Open first level when a file opened or created.
-- Fix: Flushing changes to internal file store sometimes may miss the latest
-  changes in levels.
+- Fix: Improved stability for flushing changes back to in-memory file.
+  - Fix: The latest changes in levels may be lost sometimes due to bug in logic.
+  - Fix: Flush was firing on first "page hide" event only.
+  - Enh: Reduced delay to flush changes in background from 3 to 0.3 sec.
 - Fix: Incorrect border in new MPX file with custom size.
 - Fix: Lost which levels was previously opened on page load.
 - Fix: Some text grammar.