+++ CHANGELOG.md
@@ -12,6 +12,8 @@
   - Enh: Reduced delay to flush changes in background from 3 to 0.3 sec.
 - Fix: Incorrect border in new MPX file with custom size.
 - Fix: Lost which levels was previously opened on page load.
+- Fix: Dragging pointer out of canvas while drawing cause drawing to stop and
+  breaks next one drawing.
 - Fix: Some text grammar.
 
 ## 0.1.0 (2022-10-13)